hf-deployment/requirements.txt
insightface==0.7.3
onnxruntime>=1.18.0
pillow==10.1.0
numpy==1.24.3
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
